+++ lib/transaction/view/transaction_form.dart
@@ -3,8 +3,6 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:formz/formz.dart';
 
-import '../../../constants/constants.dart';
-import '../../constants/colors.dart';
 import '../../utils/theme/cubit/theme_cubit.dart';
 import '../transaction.dart';