+++ lib/transfer/view/transfer_page.dart
@@ -33,7 +33,13 @@ class TransferView extends StatelessWidget {
       builder: (context, state) {
         return Scaffold(
             appBar: AppBar(
-              title: Text('Transfer', style: TextStyle(fontSize: 36.sp)),
+              title: Builder(builder: (context) {
+                final isIdExist =
+                    context.select((TransferBloc bloc) => bloc.state.id) !=
+                        null;
+                return Text(isIdExist ? 'Edit Transfer' : 'New Transfer',
+                    style: TextStyle(fontSize: 36.sp));
+              }),
               leading: IconButton(
                 icon: Icon(Icons.close),
                 onPressed: () {