+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 ### Added
 - New Subcategory form
 - Submit buttons appear just on the valid forms
+### Fixed
+- Transaction bar title (new/edit)
 
 ## [4.2.1] - 03-17-2025
 ### Hotfix